introduce ProjectUpdater
