+++ api/src/main/java/org/accula/api/handlers/util/package-info.java
@@ -1,0 +1,4 @@
+@NonNullApi
+package org.accula.api.handlers.util;
+
+import org.springframework.lang.NonNullApi;
